Logout feature, add logout to sidebar and rename Root.jsx to Sidebar.jsx
