+++ frontend/src/components/Sidebar.jsx
@@ -1,5 +1,6 @@
 import { Link } from "react-router-dom"
 import Styles from "../assets/css/sidebar.module.css"
+import Logout from "./Logout"
 
 export default function Root() {
     return(
@@ -10,9 +11,9 @@ export default function Root() {
                 <img src="../assets/images/logo.png" id={Styles.logo}></img>
             </Link>
             <div id={Styles.content}>
-                <p><Link to="/login">🔑 Login</Link></p>
-                <p><Link to="/profile">👤 Profile</Link></p>
-                <p><Link to="/settings">⚙️ Settings</Link></p>
+                <Link to="/profile" className={Styles.item}>👤 Profile</Link>
+                <Link to="/settings" className={Styles.item}>⚙️ Settings</Link>
+                <Logout className={Styles.item} />
             </div>
         </nav>
         </>
